backend/requirements.txt
flask
flask_cors
requests
pymongo
python-dotenv